requirements.txt
pandas
numpy
requests
xgboost
scikit-learn
matplotlib
google-generativeai
python-dotenv
mlflow